+++ shared/src/store/action-reducers/utils/log.ts
@@ -7,9 +7,9 @@ import {
     createAlarmGroupTag,
     createBehaviorTag,
     createHospitalTag,
-    createPatientStatusTag,
     createPatientTag,
     createSimulatedRegionTagWithName,
+    createPatientStatusTag,
     createRadiogramTypeTag,
     createSimulatedRegionTag,
     createTagsForSinglePatient,
@@ -18,12 +18,14 @@ import {
     createVehicleTag,
     createVehicleTypeTag,
 } from '../../../models/utils/tag-helpers';
+import type { TreatmentProgress } from '../../../simulation/utils/treatment';
 import { treatmentProgressToGermanNameDictionary } from '../../../simulation/utils/treatment';
 import type { ExerciseState } from '../../../state';
 import type { UUID } from '../../../utils';
 import { StrictObject } from '../../../utils';
 import type { Mutable } from '../../../utils/immutability';
 import { Patient } from '../../../models/patient';
+import type { PatientStatus, Personnel, Vehicle } from '../../../models';
 import {
     currentSimulatedRegionIdOf,
     currentSimulatedRegionOf,
@@ -32,7 +34,6 @@ import {
 import type { WithPosition } from '../../../models/utils/position/with-position';
 import type { TransferPoint } from '../../../models/transfer-point';
 import type { Hospital } from '../../../models/hospital';
-import type { Personnel, Vehicle } from '../../../models';
 import {
     getElement,
     getExerciseBehaviorById,
@@ -588,6 +589,96 @@ export function logRadiogram(
             ],
             state.currentTime
         )
+    );
+}
+export function logTreatmentStatusChangedInSimulatedRegion(
+    state: Mutable<ExerciseState>,
+    treatmentProgress: TreatmentProgress,
+    simulatedRegionId: UUID,
+    additionalTags: Tag[] = []
+) {
+    if (!logActive(state)) return;
+
+    const simulatedRegion = getElement(
+        state,
+        'simulatedRegion',
+        simulatedRegionId
+    );
+
+    state.logEntries!.push(
+        new LogEntry(
+            `Der Behandlungszustand im simulierten Bereich ${simulatedRegion.name} ist zu ${treatmentProgressToGermanNameDictionary[treatmentProgress]} gewechselt.`,
+            [
+                ...additionalTags,
+                createTreatmentProgressTag(state, treatmentProgress),
+                createSimulatedRegionTag(state, simulatedRegionId),
+            ],
+            state.currentTime
+        )
+    );
+}
+
+export function logLastPatientTransported(
+    state: Mutable<ExerciseState>,
+    patientStatus: PatientStatus,
+    simulatedRegionId: UUID,
+    description: string,
+    additionalTags: Tag[] = []
+) {
+    if (!logActive(state)) return;
+
+    state.logEntries!.push(
+        new LogEntry(
+            description,
+            [
+                ...additionalTags,
+                createPatientStatusTag(state, patientStatus),
+                createSimulatedRegionTag(state, simulatedRegionId),
+            ],
+            state.currentTime
+        )
+    );
+}
+
+export function logLastPatientTransportedInSimulatedRegion(
+    state: Mutable<ExerciseState>,
+    patientStatus: PatientStatus,
+    simulatedRegionId: UUID
+) {
+    if (!logActive(state)) return;
+
+    const simulatedRegion = getElement(
+        state,
+        'simulatedRegion',
+        simulatedRegionId
+    );
+
+    logLastPatientTransported(
+        state,
+        patientStatus,
+        simulatedRegionId,
+        `Der letzte Patient der Kategorie ${statusNames[patientStatus]} im simulierten Bereich ${simulatedRegion.name} ist abtransportiert worden`
+    );
+}
+
+export function logLastPatientTransportedInMultipleSimulatedRegions(
+    state: Mutable<ExerciseState>,
+    patientStatus: PatientStatus,
+    managingSimulatedRegionId: UUID
+) {
+    if (!logActive(state)) return;
+
+    const simulatedRegion = getElement(
+        state,
+        'simulatedRegion',
+        managingSimulatedRegionId
+    );
+
+    logLastPatientTransported(
+        state,
+        patientStatus,
+        managingSimulatedRegionId,
+        `Der letzte Patient der Kategorie ${statusNames[patientStatus]} in allen Bereichen, die von der TO in ${simulatedRegion.name} verwaltet werden, ist abtransportiert worden`
     );
 }